refactor: rename hmm filename
