+++ src/ss/system/__init__.py
@@ -9,7 +9,7 @@ if TYPE_CHECKING:
         simulate,
     )
     from .cart_pole import CartPoleSystem
-    from .discrete_hmm import HiddenMarkovModel
+    from .hmm import HiddenMarkovModel
     # from .linear import (
     #     ContinuousTimeLinearSystem,
     #     DiscreteTimeLinearSystem,
@@ -24,7 +24,7 @@ _EXPORTS: dict[str, str] = {
     "System": "_system",
     "simulate": "_system",
     "CartPoleSystem": "cart_pole",
-    "HiddenMarkovModel": "discrete_hmm",
+    "HiddenMarkovModel": "hmm",
     # "ContinuousTimeLinearSystem": "linear",
     # "DiscreteTimeLinearSystem": "linear",
     # "ContinuousTimeNonlinearSystem": "nonlinear",
