+++ src/ss/system/hmm.py
